+++ packages/sdk/src/interfaces/order/update-order.ts
@@ -1,7 +1,6 @@
 import type { Fcl } from "@rarible/fcl-types"
 import type { Maybe } from "@rarible/types/build/maybe"
 import type { BigNumber } from "@rarible/types"
-import { toFlowAddress } from "@rarible/types"
 import type { FlowOrder, FlowOrderControllerApi } from "@rarible/flow-api-client"
 import type { AuthWithPrivateKey, FlowCurrency, FlowNetwork } from "../../types/types"
 import { runTransaction, waitForSeal } from "../../common/transaction"
@@ -9,6 +8,7 @@ import { parseEvents } from "../../common/parse-tx-events"
 import { getOrderCode } from "../../blockchain-api/order/storefront"
 import type { FlowContractAddress } from "../../types/contract-address"
 import { getCollectionConfig } from "../../config/utils"
+import { getAccountAddress } from "../../common/get-account-address"
 import { checkPrice } from "./common/check-price"
 import { getOrderDetailsFromBlockchain } from "./common/get-order-details-from-blockchain"
 import { getPreparedOrder } from "./common/get-prepared-order"
@@ -32,10 +32,7 @@ export async function updateOrder(
 	const { collection, currency, sellItemPrice, order } = request
 	checkPrice(sellItemPrice)
 	if (fcl) {
-		const from = auth ? toFlowAddress((await auth()).addr) : toFlowAddress((await fcl.currentUser().snapshot()).addr!)
-		if (!from) {
-			throw new Error("FLOW-SDK: Can't get current user address")
-		}
+		const from = await getAccountAddress(fcl, auth)
 		const preparedOrder = await getPreparedOrder(orderApi, order)
 		const { name, map } = getCollectionConfig(network, collection)