+++ packages/sdk/src/test/check-event.ts
@@ -13,6 +13,9 @@ type EventNames =
 	| "Burn"
 	| "BidCompleted"
 	| "BidAvailable"
+	| "LotCompleted"
+	| "OpenBid"
+	| "IncreaseBid"
 
 export function checkEvent(txResult: FlowTransaction, eventName: EventNames, contractName?: FlowContractName) {
 	const result = !!txResult.events.find(e => {