+++ CHANGELOG.md
@@ -290,6 +290,18 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [0.5.30-ac.0](https://github.com/rarible/flow-sdk/compare/v0.5.28...v0.5.30-ac.0) (2021-12-30)
+
+
+### Bug Fixes
+
+* lerna config ([b1b0b1f](https://github.com/rarible/flow-sdk/commit/b1b0b1f0108c7ff339a6b0d9e4aa471b0107ca68))
+* sdk doc ([4f7bec9](https://github.com/rarible/flow-sdk/commit/4f7bec910836bb48316ea4205b19f05d3ef957f4))
+
+
+
+
+
 ## [0.5.28](https://github.com/rarible/flow-sdk/compare/v0.5.27...v0.5.28) (2021-12-21)