+++ packages/test-common/CHANGELOG.md
@@ -97,6 +97,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [0.5.30-ac.0](https://github.com/rarible/flow-sdk/compare/v0.5.28...v0.5.30-ac.0) (2021-12-30)
+
+**Note:** Version bump only for package @rarible/flow-test-common
+
+
+
+
+
 ## [0.5.14](https://github.com/rarible/flow-sdk/compare/v0.5.13...v0.5.14) (2021-12-15)